Moved to response/utils.py
