+++ package/kedro_viz/utils.py
@@ -1,11 +1,9 @@
 """Transcoding related utility functions."""
 
 import hashlib
-import logging
 import sys
 import threading
 import time
-from datetime import datetime
 from enum import Enum
 from itertools import cycle
 from pathlib import Path
@@ -14,8 +12,6 @@ from typing import Any, Dict, Optional, Tuple, TypeVar
 from pathspec import GitIgnoreSpec
 
 TRANSCODING_SEPARATOR = "@"
-
-T = TypeVar("T", bound=Enum)
 
 
 def _hash(value: str):
@@ -139,54 +135,3 @@ class Spinner:
     def stop(self):
         self.stop_running = True
         self._spinner_thread.join()
-
-
-def safe_int(value: Any) -> int:
-    """Safely parse an integer value."""
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return 0
-
-
-def convert_status_to_enum(status: Optional[str], default: T) -> T:
-    """Convert string status to enum member; case-insensitive match on values."""
-    logger = logging.getLogger(__name__)
-    if not status:
-        return default
-    enum_cls = type(default)
-    for member in enum_cls:
-        if member.value.lower() == status.lower():
-            return member
-    logger.debug("Unknown status '%s', returning default %s", status, default)
-    return default
-
-
-def calculate_pipeline_duration(
-    start_time: Optional[str],
-    end_time: Optional[str],
-    nodes_durations: Dict[str, float],
-) -> float:
-    """Calculate pipeline duration from timestamps or node durations.
-
-    Args:
-        start_time: ISO format timestamp string for pipeline start
-        end_time: ISO format timestamp string for pipeline end
-        nodes_durations: Dictionary mapping node IDs to their durations in seconds
-
-    Returns:
-        Total duration in seconds
-    """
-    logger = logging.getLogger(__name__)
-    if start_time and end_time:
-        try:
-            start_dt = datetime.fromisoformat(start_time)
-            end_dt = datetime.fromisoformat(end_time)
-            duration = (end_dt - start_dt).total_seconds()
-            logger.info(f"Duration calculated from timestamps: {duration} seconds")
-            return duration
-        except (ValueError, TypeError) as e:
-            logger.warning(f"Error calculating pipeline duration: {e}")
-
-    # Fallback to summing up node durations
-    return sum(nodes_durations.values())
